+++ src/no/turan/live/android/sensors/IHRSensor.java
@@ -1,5 +1,5 @@
 package no.turan.live.android.sensors;
 
 public interface IHRSensor extends ISensor {
-	public short getHR();
+	public int getHR();
 }
